backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "asyncpg>=0.30.0",
    "beautifulsoup4>=4.14.3",
    "fastapi>=0.129.0",
    "groq>=1.0.0",
    "langchain-text-splitters>=0.3.0",
    "lxml>=5.3.0",
    "pdfplumber>=0.11.9",
    "pydantic-settings>=2.13.0",
    "python-docx>=1.2.0",
    "python-dotenv>=1.2.1",
    "python-multipart>=0.0.22",
    "requests>=2.32.5",
    "tavily-python>=0.7.21",
    "uvicorn>=0.41.0",
]

[dependency-groups]
dev = [
    "httpx>=0.27.0",
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
]
